+++ modules/versions.lua
@@ -3,7 +3,7 @@
 -- mod.find. Do not edit; rebuild.
 return {
   ["Gen1Arena"] = "0.18.3",
-  ["Gen1BattleUI"] = "1.4.0",
+  ["Gen1BattleUI"] = "1.5.0",
   ["Gen1BillsBox"] = "1.2.0",
   ["Gen1Dex"] = "1.5.0",
   ["Gen1MenuManager"] = "0.2.6",